[WIP] lib/miq_seeds/generator/vm_infra_explorer.rb
Honestly, not sure where I was going with this one.  This is like a
year+ old at this point, so not sure it is worth keeping.  Pushing it up
though incase I find a use for it.

Going to be refactoring this project anyway though, as I think this can
be done much, and made much more flexible.